waylandsink: Wait until a wl_surface_frame event has been received before dropping frames.
The wl_display events dispatching runs asynchronously in a seperate
thread from the streaming thread. This means that processing a
frame callback event could be slightly late, causing unnecessary
frame drops when the video framerate is the almost same as a display
vsync frequency.

This change solves this problem by dispatching any pending events
in the streaming thread.