gstwaylandsink: WORKAROUND: Clear final output frame when moving from PAUSED to READY
This workaround will stop the "pages still in use" kernel warning from
appearing at the end of playback when gst-omx and waylandsink are connected
directly AND gst-omx manages the buffer pool.

This patch pushes a blank frame to the screen and waits for
the last frame to be released by the compositor before moving to the READY
state, where gst-omx will force free all of the output frame memory.
The side effect of this patch is that there is no more fadeout effect at the
end of playback, since the final frame has been removed from the compositor.