waylandsink: Don't try to re-render a buffer that is already onscreen.
Re-submitting an already onscreen buffer to the compositor will increase the
reference count on the GstBuffer, but (with a Weston compositor) will not
produce a second buffer release event.  As a result there will only be
one unref and two refs on the GstBuffer, preventing it from ever returning
to the buffer pool.  A double rendering attempt will occur on the first frame
after prerolling, since the same frame is submitted from both the preroll and
the playing handlers.