waylandsink: Don't close the wl_display until wayland buffers are freed
When using buffers from an external buffer pool, the local
(waylandsink internal) pool does not contain any buffers and so can be freed
even when buffers are still in use in the wayland server. The local pool
holds the only referece to the wayland display, so freeing the pool will
close the display.  This patch gives each buffer a reference to the display
so that is will not be closed while the buffers are still in use.